requirements.txt
Django==3.1.6
djangorestframework==3.12.2
django-filter==2.4.0
Markdown==3.3.3
djangorestframework-guardian==0.3.0
django-storages==1.11.1
